requirements.txt
absl-py==2.1.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.34.159
botocore==1.34.159
CacheMan==2.2.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
cyhunspell==2.0.2
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
enchant==0.0.1
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.4.0
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.65.4
h11==0.14.0
h5py==3.11.0
hdfviewer==0.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
importlib-metadata==8.2.0
importlib-resources==6.4.0
ipykernel==6.29.5
ipympl==0.9.3
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.3
jedi==0.19.1
jinja2==3.1.4
jmespath==1.0.1
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.2
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.2.4
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.11
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook-shim==0.2.4
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.2
prometheus-client==0.20.0
prompt-toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
pure-eval==0.2.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-magic==0.4.27
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.1.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
scipy==1.10.1
selenium==4.23.1
Send2Trash==1.8.3
six==1.16.0
skipy==0.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.1
tinycss2==1.3.0
tk==0.1.0
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
wsproto==1.2.0
zipp==3.20.0
streamlit>=1.30
streamlit-webrtc>=0.47


